Add `output_file` option to save scan results as a file (#53)
Currently, the JSON results are uploaded as an artifact by default, making them accessible to other jobs, but also requiring a download action if post-processing is desired.

Users may wish to save the scan results -- regardless of requested format -- to a file in the workspace instead, so it is easily accessible to next steps. Setting this option will specify a file name.

The file name is also available in `steps.<id>.outputs.results` when this option is set.